requirements.txt
Flask==2.3.2
opencv-python-headless==4.8.1.78
numpy==1.23.5
Pillow==10.0.0